ci(yomu): diff-cover で test ファイルをカバレッジ計測から除外
cargo llvm-cov の --ignore-filename-regex に /tests\.rs$ を追加し、全 *.rs の split-out test モジュールを diff coverage の計測対象から外す。リファクタで test ファイルを移動すると to-fail panic アーム (成功時は到達しない other => panic! 等) が changed-lines に入り未カバー判定で 95% ゲートを落とす問題 (#257 / #259) を恒久解消する。

THRESHOLDS.md の「test files はカバレッジ例外」を ci.yml 側に実装し、既に test を除外する file-size gate との非対称も解消する。本番コード (インライン #[cfg(test)] を含む本体ファイル) の計測は維持する。

あわせて、最後に残ったインラインテスト src/cli.rs の mod tests を src/cli/tests.rs へ分離 (#251 のインラインテスト分離方針の完了)。

Closes #260
