+++ .github/workflows/ci.yml
@@ -140,8 +140,13 @@ jobs:
         # cfg'd out here. `model download` (real network I/O) and `emit_query_log`
         # (real FS + env I/O; its logic is unit-tested in query_log) are reachable
         # from production dispatch and cannot be cfg'd out, so their files are
-        # excluded by path — like the real-model loader.
-        run: cargo llvm-cov -p yomu --lcov --output-path lcov.info --features test-support --ignore-filename-regex 'tools/(model|query_emit)\.rs'
+        # excluded by path — like the real-model loader. Split-out `*/tests.rs`
+        # modules are excluded too (THRESHOLDS exempts tests; mirrors the
+        # file-size gate above): a refactor that moves a test file would otherwise
+        # make its to-fail panic arms changed lines that diff-cover scores
+        # uncovered (#257, #259, #260). Production bodies (incl. inline
+        # `#[cfg(test)]`) stay measured.
+        run: cargo llvm-cov -p yomu --lcov --output-path lcov.info --features test-support --ignore-filename-regex 'tools/(model|query_emit)\.rs|/tests\.rs$'
 
       - name: Diff coverage gate (changed lines >= 95%)
         run: |
